+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## Misc
+
+* enable `jinja2.autoescape` for HTML/XML templates (ref: https://jinja.palletsprojects.com/en/stable/api/#autoescaping)
+
 ## 0.23.1 (2025-08-27)
 
 ### titiler.core